creating a parent class: BidsNidm
