+++ nidm/experiment/__init__.py
@@ -8,4 +8,5 @@ from .AcquisitionObject import AcquisitionObject
 from .MRObject import MRObject
 from .DemographicsObject import DemographicsObject
 from .AssessmentObject import AssessmentObject
-from .bidsmri import BidsMriProject
+from .bids_base import BidsNidm
+from .bidsmri import BidsMriNidm
